Abstracted class setup.
