+++ test/test_catalog.py
@@ -4,22 +4,10 @@ import unittest
 
 import requests.auth
 
+from test_broker_base import TestBrokerBase
 
-class TestCatalog(unittest.TestCase):
-    def setUp(self):
-        os.environ['BROKER_HOST'] = str("http://localhost:8090")
-        os.environ['BROKER_USERNAME'] = str("admin")
-        os.environ['BROKER_PASSWORD'] = str("password")
 
-        self.host = os.environ['BROKER_HOST']
-        self.username = os.environ['BROKER_USERNAME']
-        self.password = os.environ['BROKER_PASSWORD']
-
-        self.auth = requests.auth.HTTPBasicAuth(self.username, self.password)
-        self.headers = {
-            "X-Broker-API-Version": "2.14",
-        }
-
+class TestCatalog(TestBrokerBase, unittest.TestCase):
     def test_catalog(self):
         url = self.host + "/v2/catalog"
         r = requests.get(url, auth=self.auth, headers=self.headers)
