+++ test/test_provision.py
@@ -8,27 +8,10 @@ import uuid
 
 import requests.auth
 
+from test_broker_base import TestBrokerBase
 
-class TestProvision(unittest.TestCase):
-    @classmethod
-    def setUpClass(self):
-        os.environ['BROKER_HOST'] = str("http://localhost:8090")
-        os.environ['BROKER_USERNAME'] = str("admin")
-        os.environ['BROKER_PASSWORD'] = str("password")
 
-        self.host = os.environ['BROKER_HOST']
-        self.username = os.environ['BROKER_USERNAME']
-        self.password = os.environ['BROKER_PASSWORD']
-
-        self.auth = requests.auth.HTTPBasicAuth(self.username, self.password)
-        self.headers = {
-            "X-Broker-API-Version": "2.14",
-        }
-
-        self.instance_id = uuid.uuid4()
-        self.service_id = "c76ed0a4-9a04-5710-90c2-75e955697b08"
-        self.plan_id = self.service_id + "-small"
-
+class TestProvision(TestBrokerBase, unittest.TestCase):
     def test_provision(self):
         url = self.host + "/v2/service_instances/{}?accepts_incomplete=true".format(self.instance_id)
         r = requests.put(url, auth=self.auth, headers=self.headers, data=json.dumps({
@@ -41,6 +24,7 @@ class TestProvision(unittest.TestCase):
         diff = datetime.timedelta(seconds=0)
         state = "in progress"
         while state == "in progress" and diff < datetime.timedelta(minutes = 2):
+            print("provisioning in progress, state: {}...".format(state))
             time.sleep(5)
             url = self.host + "/v2/service_instances/{}/last_operation?operation=provision&service_id={}&plan_id={}".format(
                 self.instance_id, self.service_id, self.plan_id
@@ -52,7 +36,6 @@ class TestProvision(unittest.TestCase):
             self.assertEqual(200, r.status_code)
             json_body = json.loads(r.content.decode())
             state = json_body["state"]
-            print("provisioning in progress, state: {}...".format(state))
             now_time = datetime.datetime.now()
             diff = now_time - start_time
